pyproject.toml
[project]
name = "newsstand-back"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-stubs>=3.16.9",
    "Pillow>=12.1.1",
    "pydantic>=2.12.5",
    "python-decouple>=3.8",
]
